added jwt authorization token for deleting post in api v1

+++ controllers/api/v1/posts_api.js
@@ -23,13 +23,18 @@ module.exports.index = async (req,res)=>{
 module.exports.destroy = async(req,res)=>{
     try{
         const post = await Post.findById(req.params.id);
+        if(post.user == req.user.id){
             await post.deleteOne();
             await Comment.deleteMany({post : req.params.id});
 
             return res.status(200).json({
                 message : 'post and associated comments deleted'
             })
-        
+        }else{
+            return res.status(401).json({
+                message : "You cannot delete this post"
+            })
+        }
     }catch(err){
         req.flash('error',err)
         return res.status(500).json({message:'Error in deleting'})
